MNMS-454 Feat&Chore&Refactor : 지인 양도 코드 추가 및 ocr 임시 수정 및 리팩토링
지인 양도 코드 추가 : 결제 kafka 구독 (테스트 미완)
양도 리팩토링 진행 중
ocr 임시 수정 : ocr parser 문제 임시 수정 (조건 느슨하게 수정)
