+++ src/main/java/com/mnms/booking/dto/request/TicketTransferRequestDTO.java
@@ -8,6 +8,4 @@ public class TicketTransferRequestDTO {
     private Long recipientId; // 양도받을 가족/지인 ID
     private String transferType; // "FAMILY" 또는 "OTHERS"
     private String senderName;
-    //private String receiveMethod; // 수령 방법 (예: QR, 배송)
-    //private String address;       // 필요시
 }
