+++ src/main/java/com/mnms/booking/dto/response/TransferOthersResponseDTO.java
@@ -1,5 +1,8 @@
 package com.mnms.booking.dto.response;
 
+import com.mnms.booking.entity.Festival;
+import com.mnms.booking.entity.Ticket;
+import com.mnms.booking.entity.Transfer;
 import lombok.Builder;
 import lombok.Data;
 
@@ -21,4 +24,17 @@ public class TransferOthersResponseDTO {
     private int ticketPrice;
     private String fname;
     private String posterFile;
+
+    public static TransferOthersResponseDTO from(Transfer transfer, Ticket ticket, Festival festival, Long userId) {
+        return TransferOthersResponseDTO.builder()
+                .reservationNumber(ticket.getReservationNumber())
+                .senderId(transfer.getSenderId())
+                .receiverId(userId)
+                .selectedTicketCount(ticket.getSelectedTicketCount())
+                .ticketPrice(festival.getTicketPrice())
+                .fname(festival.getFname())
+                .posterFile(festival.getPosterFile())
+                .performanceDate(ticket.getPerformanceDate())
+                .build();
+    }
 }
