+++ src/main/java/com/mnms/booking/kafka/listener/PaymentListener.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.mnms.booking.kafka.dto.PaymentSuccessEventDTO;
 import com.mnms.booking.service.BookingCommandService;
+import com.mnms.booking.service.TransferCompletionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -17,6 +18,7 @@ import org.springframework.stereotype.Service;
 public class PaymentListener {
 
     private final BookingCommandService bookingCommandService;
+    private final TransferCompletionService transferCompletionService;
 
     @KafkaListener(topics = "${app.kafka.topic.payment-event}", groupId = "booking-service-group")
     public void consumePaymentSuccess(String message) throws JsonProcessingException {
@@ -38,5 +40,16 @@ public class PaymentListener {
 
         PaymentSuccessEventDTO event = objectMapper.readValue(message, PaymentSuccessEventDTO.class);
         bookingCommandService.cancelBooking(event.getReservationNumber(), event.isSuccess());
+    }
+
+    @KafkaListener(topics = "${app.kafka.topic.transfer-payment-event}", groupId = "booking-service-group")
+    public void consumeTransferSuccess(String message) throws JsonProcessingException {
+        ObjectMapper objectMapper = new ObjectMapper();
+
+        objectMapper.registerModule(new JavaTimeModule());
+        objectMapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
+
+        PaymentSuccessEventDTO event = objectMapper.readValue(message, PaymentSuccessEventDTO.class);
+        transferCompletionService.updateOthersTicket(event.getReservationNumber(), event.isSuccess());
     }
 }
