+++ src/main/java/com/mnms/booking/repository/TransferRepository.java
@@ -1,5 +1,6 @@
 package com.mnms.booking.repository;
 
+import com.mnms.booking.entity.Ticket;
 import com.mnms.booking.entity.Transfer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -7,6 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface TransferRepository extends JpaRepository<Transfer, Long> {
+    Transfer findByTicket(Ticket ticket);
     List<Transfer> findByReceiverId(Long userId);
     Optional<Transfer> findById(Long id);
 }
