+++ src/main/java/com/mnms/booking/service/TransferCompletionService.java
@@ -1,0 +1,194 @@
+package com.mnms.booking.service;
+
+import com.mnms.booking.dto.request.UpdateTicketRequestDTO;
+import com.mnms.booking.dto.response.TransferOthersResponseDTO;
+import com.mnms.booking.entity.Festival;
+import com.mnms.booking.entity.QrCode;
+import com.mnms.booking.entity.Ticket;
+import com.mnms.booking.entity.Transfer;
+import com.mnms.booking.enums.TicketType;
+import com.mnms.booking.enums.TransferStatus;
+import com.mnms.booking.enums.TransferType;
+import com.mnms.booking.exception.BusinessException;
+import com.mnms.booking.exception.ErrorCode;
+import com.mnms.booking.repository.QrCodeRepository;
+import com.mnms.booking.repository.TicketRepository;
+import com.mnms.booking.repository.TransferRepository;
+import com.mnms.booking.util.CommonUtils;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDate;
+import java.util.List;
+
+///  양도 수락
+@Service
+@Slf4j
+@RequiredArgsConstructor
+@Transactional
+public class TransferCompletionService {
+
+    private final TicketRepository ticketRepository;
+    private final QrCodeRepository qrCodeRepository;
+    private final TransferRepository transferRepository;
+    private final CommonUtils commonUtils;
+    private final QrCodeService qrCodeService;
+
+    /// 가족 간 양도 수락
+    @Transactional(rollbackFor = BusinessException.class)
+    public void updateFamilyTicket(UpdateTicketRequestDTO request, Long userId) {
+        Transfer transfer = getTransferOrThrow(request.getTransferId());
+
+        validateSenderId(request.getSenderId(), transfer.getTicket().getUserId());
+        validateTransferType(transfer, TransferType.FAMILY);
+        validateReceiver(transfer, userId);
+
+        // CANCELED 처리
+        if (request.getTransferStatus() == TransferStatus.CANCELED) {
+            transfer.setStatus(TransferStatus.CANCELED);
+            return;
+        }
+
+        // ticket 점검
+        Ticket ticket = getTicketOrThrow(transfer.getTicket().getReservationNumber());
+        validateTicketStatus(ticket);
+
+        // transfer 상태 업데이트
+        transfer.setStatus(TransferStatus.COMPLETED);
+        transfer.setTicketType(request.getTicketType());
+        transfer.setAddress(request.getAddress());
+
+        // ticket, QR 정보 업데이트
+        updateTicketInfo(ticket, request, userId);
+        updateQrCodes(ticket, request, userId);
+    }
+
+    /// 지인 간 양도 요청 수락
+    @Transactional
+    public TransferOthersResponseDTO proceedOthersTicket(UpdateTicketRequestDTO request, Long userId) {
+
+        Transfer transfer = getTransferOrThrow(request.getTransferId());
+        validateSenderId(request.getSenderId(), transfer.getTicket().getUserId());
+        validateTransferType(transfer, TransferType.OTHERS);
+
+        String reservationNumber = transfer.getTicket().getReservationNumber();
+        validateReceiver(transfer, userId);
+
+        // CANCELED 처리
+        if (request.getTransferStatus() == TransferStatus.CANCELED) {
+            transfer.setStatus(TransferStatus.CANCELED);
+            return TransferOthersResponseDTO.builder()
+                    .reservationNumber(reservationNumber)
+                    .senderId(transfer.getSenderId())
+                    .receiverId(userId)
+                    .build();
+        }
+
+        Ticket ticket = getTicketOrThrow(reservationNumber);
+        validateTicketStatus(ticket);
+        Festival festival = ticket.getFestival();
+
+        // transfer 업데이트
+        transfer.setStatus(TransferStatus.APPROVED);
+        transfer.setTicketType(request.getTicketType());
+        transfer.setAddress(request.getAddress());
+
+        return TransferOthersResponseDTO.from(transfer, ticket, festival, userId);
+    }
+
+
+    /// 결제 KAFKA 구독 메시지 받고 결제 완료 수행
+    @Transactional
+    public void updateOthersTicket(String reservationNumber, boolean paymentStatus) {
+        Ticket ticket = ticketRepository.findByReservationNumber(reservationNumber)
+                .orElseThrow(() -> new BusinessException(ErrorCode.TICKET_NOT_FOUND));
+
+        Transfer transfer = transferRepository.findByTicket(ticket);
+        if (!paymentStatus) {
+            transfer.setStatus(TransferStatus.CANCELED);
+            return;
+        }
+
+        transfer.setStatus(TransferStatus.COMPLETED);
+
+        UpdateTicketRequestDTO request = UpdateTicketRequestDTO.builder()
+                .transferId(transfer.getId())
+                .senderId(transfer.getSenderId())
+                .transferStatus(transfer.getStatus())
+                .ticketType(transfer.getTicketType())
+                .address(transfer.getAddress())
+                .build();
+
+        updateTicketInfo(ticket, request, transfer.getReceiverId());
+        updateQrCodes(ticket, request, transfer.getReceiverId());
+    }
+
+
+    /// UTIL
+    private void validateTransferType(Transfer transfer, TransferType expectedType) {
+        if (!transfer.getType().equals(expectedType)) {
+            throw new BusinessException(ErrorCode.TRANSFER_NOT_MATCH_TYPE);
+        }
+    }
+
+    private void validateSenderId(Long senderId, Long ticketUserId) {
+        if (!senderId.equals(ticketUserId)) {
+            throw new BusinessException(ErrorCode.TRANSFER_NOT_MATCH_SENDER);
+        }
+    }
+
+    private void validateReceiver(Transfer transfer, Long userId){
+        // 양수자 확인
+        if (!userId.equals(transfer.getReceiverId())){
+            throw new BusinessException(ErrorCode.TRANSFER_NOT_MATCH_RECEIVER);
+        }
+    }
+    private Transfer getTransferOrThrow(Long transferId) {
+        return transferRepository.findById(transferId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.TRANSFER_NOT_EXIST));
+    }
+
+    private Ticket getTicketOrThrow(String reservationNumber) {
+        return ticketRepository.findByReservationNumber(reservationNumber)
+                .orElseThrow(() -> new BusinessException(ErrorCode.TICKET_NOT_FOUND));
+    }
+
+    private void validateTicketStatus(Ticket ticket) {
+        if (ticket.isExpired()) {
+            throw new BusinessException(ErrorCode.TICKET_EXPIRED);
+        }
+        if (ticket.isCanceled()) {
+            throw new BusinessException(ErrorCode.TICKET_CANCELED);
+        }
+    }
+
+    /// ticket 정보 업데이트
+    private void updateTicketInfo(Ticket ticket, UpdateTicketRequestDTO request, Long receiverId) {
+        TicketType deliveryMethod = request.getTicketType();
+
+        ticket.updateTicketInfo(
+                commonUtils.generateReservationNumber(),
+                deliveryMethod,
+                receiverId,
+                LocalDate.now(),
+                request.getAddress()
+        );
+    }
+
+    /// qr 정보 업데이트
+    private void updateQrCodes(Ticket ticket, UpdateTicketRequestDTO request,  Long receiverId) {
+        List<QrCode> existingQrs = qrCodeRepository.findByTicketId(ticket.getId())
+                .orElseThrow(() -> new BusinessException(ErrorCode.QR_CODE_NOT_FOUND));
+
+        existingQrs.forEach(qr -> {
+            qr.setQrCodeId(qrCodeService.generateQrCodeId());
+            qr.setUserId(receiverId);
+            qr.setTicket(ticket);
+        });
+
+        ticket.getQrCodes().clear();
+        ticket.getQrCodes().addAll(existingQrs);
+    }
+}
