+++ src/main/resources/application-dev.properties
@@ -28,6 +28,7 @@ spring.data.redis.port=${REDIS_PORT}
 app.kafka.topic.festival-event=festival-topic
 app.kafka.topic.payment-event= payment-status-events
 app.kafka.topic.payment-cancel-event=payment-cancel-events
+app.kafka.topic.transfer-payment-event=payment-transfer-events
 
 spring.kafka.bootstrap-servers=${KAFKA_SERVERS}
 spring.kafka.consumer.auto-offset-reset=earliest
